requirements.txt
tornado==5.1
numpy==1.15.0
redis==2.10.6
psycopg2_binary==2.7.5
aiopg==0.15.0
coils==1.0.9
python_bcrypt==0.3.2
